### 2016-07-10 #37 @whedon
Submission: tidytext: Text Mining and Analysis Using Tidy Data Principles in R
**Submitting author:** @juliasilge (<a href="http://orcid.org/0000-0002-3671-836X">Julia Silge</a>)
**Repository:** <a href="https://github.com/juliasilge/tidytext" target ="_blank">https://github.com/juliasilge/tidytext</a>
**Version:** v0.1.1
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.56714" target ="_blank">http://dx.doi.org/10.5281/zenodo.56714</a>
