### 2016-07-11 #37 @rmflight
Regarding LICENSE, the LICENSE is the standard CRAN implementation of the MIT LICENSE, which while not exactly standard for OSS, is the standard for CRAN. Therefore I am checking it off.
