### 2016-07-11 #37 @rmflight
Below is my review, which is basically copy pasting the above checklist, with some commentary interspersed. All checked off, I recommend to **publish as-is**.

@juliasilge and @dgrtwo: one of the vignettes installing from CRAN did not have the correct title, but I see in the github repo it has been corrected, so no issues. That was the only beef I had.

> Regarding LICENSE, the LICENSE is the standard CRAN implementation of the MIT LICENSE, which while not exactly standard for OSS, is the standard for CRAN. Therefore I am checking it off.

> dependencies are handled by the R package installation

> The README provides well worked examples showing how the software works, on a real dataset

> Various functions have examples, and also included are 4 vignettes (R long-form documentation) that also demonstrate various analyses.

## Issues with Submission

The CRAN installed version does not have the correct title for one of the vignettes, but I would think that it will propagate soon.

## Recommendation

**Publish as-is**
