### 2016-07-11 #37 @juliasilge
Regarding the package's LICENSE, I did try submitting to CRAN with the contents of the whole MIT license in the LICENSE file, as discussed in the [author guidelines](http://joss.theoj.org/about#author_guidelines), but the CRAN administrators did not go for it.
