### 2016-07-11 #37 @rmflight
Right, and I figured that was an issue. That is why I just checked it anyway, because you followed the CRAN guidelines, which is where the package is "published" for ease of installation purposes.
