### 2016-07-11 f675090
Add JOSS badge to README
