+++ README.md
@@ -11,6 +11,7 @@ tidytext: Text mining using dplyr, ggplot2, and other tidy tools
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/tidytext)](https://cran.r-project.org/package=tidytext)
 [![Coverage Status](https://img.shields.io/codecov/c/github/juliasilge/tidytext/master.svg)](https://codecov.io/github/juliasilge/tidytext?branch=master)
 [![DOI](https://zenodo.org/badge/22224/juliasilge/tidytext.svg)](https://zenodo.org/badge/latestdoi/22224/juliasilge/tidytext)
+[![status](http://joss.theoj.org/papers/89fd1099620268fe0342ffdcdf66776f/status.svg)](http://joss.theoj.org/papers/89fd1099620268fe0342ffdcdf66776f)
 
 
 
@@ -51,8 +52,7 @@ original_books <- austen_books() %>%
   ungroup()
 
 original_books
-#> Source: local data frame [73,422 x 3]
-#> 
+#> # A tibble: 73,422 x 3
 #>                     text                book linenumber
 #>                    <chr>              <fctr>      <int>
 #> 1  SENSE AND SENSIBILITY Sense & Sensibility          1
@@ -65,7 +65,7 @@ original_books
 #> 8                        Sense & Sensibility          8
 #> 9                        Sense & Sensibility          9
 #> 10             CHAPTER 1 Sense & Sensibility         10
-#> ..                   ...                 ...        ...
+#> # ... with 73,412 more rows
 ```
 
 To work with this as a tidy dataset, we need to restructure it as **one-token-per-row** format. The `unnest_tokens` function is a way to convert a dataframe with a text column to be one-token-per-row:
@@ -77,8 +77,7 @@ tidy_books <- original_books %>%
   unnest_tokens(word, text)
 
 tidy_books
-#> Source: local data frame [725,054 x 3]
-#> 
+#> # A tibble: 725,054 x 3
 #>                   book linenumber        word
 #>                 <fctr>      <int>       <chr>
 #> 1  Sense & Sensibility          1       sense
@@ -91,10 +90,10 @@ tidy_books
 #> 8  Sense & Sensibility         10     chapter
 #> 9  Sense & Sensibility         10           1
 #> 10 Sense & Sensibility         13         the
-#> ..                 ...        ...         ...
+#> # ... with 725,044 more rows
 ```
 
-This function uses the [tokenizers package](https://github.com/lmullen/tokenizers) to separate each line into words. The default tokenizing is for words, but other options include characters, ngrams, sentences, lines, paragraphs, or separation around a regex pattern.
+This function uses the [tokenizers package](https://github.com/lmullen/tokenizers) to separate each line into words. The default tokenizing is for words, but other options include characters, n-grams, sentences, lines, paragraphs, or separation around a regex pattern.
 
 Now that the data is in one-word-per-row format, we can manipulate it with tidy tools like dplyr. We can remove stop words (kept in the tidytext dataset `stop_words`) with an `anti_join`.
 
@@ -111,8 +110,7 @@ We can also use `count` to find the most common words in all the books as a whol
 ```r
 tidy_books %>%
   count(word, sort = TRUE) 
-#> Source: local data frame [13,914 x 2]
-#> 
+#> # A tibble: 13,914 x 2
 #>      word     n
 #>     <chr> <int>
 #> 1    miss  1855
@@ -125,7 +123,7 @@ tidy_books %>%
 #> 8    emma   787
 #> 9  sister   727
 #> 10  house   699
-#> ..    ...   ...
[…]
