### 2016-07-11 c3e8031
Moved creation of "books" variable in vignette into inst/books.rda. This is to ensure the vignette can always be built. See the following issue, where Project Gutenberg sometimes can't be accessed from Travis/Appveyor:
https://github.com/ropenscilabs/gutenbergr/issues/6
