### 2016-07-11 624ad17
Oops- added back in printing of "books" variable
