### 2016-07-12 #37 @arfon
Thanks for the rapid review @rmflight!

@juliasilge - the final thing here is to make an archive of the software in Zenodo/figshare/other similar service and update this thread with the archive DOI. I can then move forward with accepting this submission :grin:
