### 2016-07-12 #37 @juliasilge
I turned on Zenodo before our last, most recent release so there is an archive of v0.1.1, the current version and what is on CRAN right now:
[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.56714.svg)](http://dx.doi.org/10.5281/zenodo.56714)

Let me know if anything needs to be changed about that, or if I need to make another release for the JOSS submission specifically or anything like that.
