### 2016-07-12 #37 @arfon
@juliasilge - this is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00037 🎉 🚀 💥
